Render collapse labels statically
Previously, both the labels and their corresponding collapse data were
being fetched and rendered dynamically. This made it so that the labels
would not be rendered until the fetch request was complete. Although
the labels would not flash when rendered, from a UX perspective it just
seemed a little "strange".

Because we know the two filter collapses types (Subreddits and Tags),
this change makes it so that the labels themselves are rendered at the
same time as the rest of the layout while their collapse data is still
fetched dynamically. Since the collapses are closed by default to begin
with, this makes for a much nicer UX.