Fix bug causing excessive network requests and slow renders
The initial `currentFilter` state would be set to an empty object upon
mounting. This meant that if a page other than Home was loaded or
refreshed, two requests to the `/v1/things` endpoint would be triggered:
one for the intial empty current filter, and another for the current
filter that is set by the page.

This in turn would cause the ThingsProvider to update the `things` state
multiple times, causing the app to be really slow when loaded or refreshed.

By using a function to check the pathname and correctly set the initial
current filter, the above issue is mitigated and making the app much more
snappier upon loading/refreshing.