Bugfix: Show spinner and disable tag input when saving tags (#12)
Fixes #11. Minor improvement to the UX wherein saving tags displays a
saving indicator and the tag input field becomes disabled. This is to
prevent users from trying to add more tags when the update tags API
request is already in-flight.

A hint/tip message was also added to teach users that multiple tags can
be added before needing to save.